+++ src/test/run-pass/warn-ctypes-inhibit.rs
@@ -1,20 +1,13 @@
-// Copyright 2012-2014 The Rust Project Developers. See the COPYRIGHT
-// file at the top-level directory of this distribution and at
-// http://rust-lang.org/COPYRIGHT.
-//
-// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
-// http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
-// <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
-// option. This file may not be copied, modified, or distributed
-// except according to those terms.
+#![allow(dead_code)]
+// compile-flags:-D improper-ctypes
 
-// compile-flags:-D ctypes
+// pretty-expanded FIXME #23616
 
-#![allow(ctypes)]
+#![allow(improper_ctypes)]
 
 mod libc {
     extern {
-        pub fn malloc(size: int) -> *const u8;
+        pub fn malloc(size: isize) -> *const u8;
     }
 }
